+++ install-stubs/Orchid/Screens/Examples/ExampleChartsScreen.php
@@ -1,0 +1,75 @@
+<?php
+
+namespace App\Orchid\Screens\Examples;
+
+use App\Orchid\Layouts\Examples\ChartBarExample;
+use App\Orchid\Layouts\Examples\ChartLineExample;
+use App\Orchid\Layouts\Examples\ChartPercentageExample;
+use App\Orchid\Layouts\Examples\ChartPieExample;
+use Orchid\Screen\Screen;
+
+class ExampleChartsScreen extends Screen
+{
+    /**
+     * Display header name.
+     *
+     * @var string
+     */
+    public $name = 'Charts';
+
+    /**
+     * Query data.
+     *
+     * @return array
+     */
+    public function query(): array
+    {
+        return [
+            'charts' => [
+                [
+                    'name'   => 'Some Data',
+                    'values' => [25, 40, 30, 35, 8, 52, 17],
+                ],
+                [
+                    'name'   => 'Another Set',
+                    'values' => [25, 50, -10, 15, 18, 32, 27],
+                ],
+                [
+                    'name'   => 'Yet Another',
+                    'values' => [15, 20, -3, -15, 58, 12, -17],
+                ],
+                [
+                    'name'   => 'And Last',
+                    'values' => [10, 33, -8, -3, 70, 20, -34],
+                ],
+            ],
+        ];
+    }
+
+    /**
+     * Button commands.
+     *
+     * @return Action[]
+     */
+    public function commandBar(): array
+    {
+        return [];
+    }
+
+    /**
+     * Views.
+     *
+     * @throws \Throwable
+     *
+     * @return array
+     */
+    public function layout(): array
+    {
+        return [
+            ChartLineExample::class,
+            ChartBarExample::class,
+            ChartPercentageExample::class,
+            ChartPieExample::class,
+        ];
+    }
+}