+++ resources/views/dummy/block.blade.php
@@ -1,4 +1,4 @@
-<div class="w-100 border v-center"
-     style="min-height: 250px;background-color: #{{ random_int(0,7). random_int(3,6) . random_int(1,9) }}">
-    <h2 class="text-muted center font-thin label">Dummy</h2>
+<div class="w-100 border v-center border-dashed my-3"
+     style="min-height: 250px;">
+    <h2 class="text-muted center font-thin">Dummy</h2>
 </div>