+++ resources/views/layouts/columns.blade.php
@@ -1,7 +1,7 @@
 <div class="hbox hbox-auto-xs">
     @foreach($manyForms as $key => $column)
         <div class="hbox-col">
-            <div class="vbox py-3  @if ($loop->first) pl-0 @endif @if ($loop->last) pr-0 @endif">
+            <div class="vbox @if(!$loop->first) pl-md-2 @endif @if(!$loop->last) pr-md-2 @endif">
                 @foreach($column as $item)
                     {!! $item ?? '' !!}
                 @endforeach