+++ resources/views/layouts/modal.blade.php
@@ -38,17 +38,21 @@
                     </div>
                     <div class="modal-footer">
 
-                        <button type="button" class="btn btn-link" data-dismiss="modal">
-                            {{ $close }}
-                        </button>
+                        @if(!$withoutCloseButton)
+                            <button type="button" class="btn btn-link" data-dismiss="modal">
+                                {{ $close }}
+                            </button>
+                        @endif
 
                         @empty($commandBar)
-                            <button type="submit"
-                                    id="submit-modal-{{$key}}"
-                                    data-turbolinks="{{ var_export($turbolinks) }}"
-                                    class="btn btn-default">
-                                {{ $apply }}
-                            </button>
+                            @if(!$withoutApplyButton)
+                                <button type="submit"
+                                        id="submit-modal-{{$key}}"
+                                        data-turbolinks="{{ var_export($turbolinks) }}"
+                                        class="btn btn-default">
+                                    {{ $apply }}
+                                </button>
+                            @endif
                         @else
                             {!! $commandBar !!}
                         @endempty