+++ src/Screen/Fields/Code.php
@@ -69,9 +69,7 @@ class Code extends Field
      * @var array
      */
     protected $inlineAttributes = [
-        'accept',
         'accesskey',
-        'autocomplete',
         'autofocus',
         'checked',
         'disabled',
@@ -83,21 +81,11 @@ class Code extends Field
         'formtarget',
         'language',
         'lineNumbers',
-        'list',
-        'max',
-        'maxlength',
-        'min',
-        'multiple',
         'name',
-        'pattern',
         'placeholder',
         'readonly',
         'required',
-        'size',
-        'src',
-        'step',
         'tabindex',
-        'type',
         'value',
         'height',
     ];